requirements.txt
asgiref==3.2.10
backcall==0.1.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
boto3==1.11.8
botocore==1.14.17
certifi==2021.10.8
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.3
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.13.0
django-decouple==2.1
django-filter==2.3.0
django-heroku==0.3.1
django-import-export==2.7.1
django-rest-framework-mongoengine==3.4.1
django-storages==1.10
djangoflutterwave==0.3.0
djangorestframework==3.13.1
djongo==1.3.3
docutils==0.15.2
et-xmlfile==1.1.0
filelock==3.4.0
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.8
importlib-metadata==1.3.0
Jinja2==3.0.3
jmespath==0.9.5
MarkupPy==1.14
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
numpy==1.17.4
odfpy==1.4.1
opencv-python==4.4.0.42
openpyxl==3.0.9
pandas==0.25.3
pbr==5.8.0
Pillow==7.1.2
platformdirs==2.4.0
psycopg2==2.9.3
pymongo==4.0.1
python-dateutil==2.8.2
pytz==2019.3
PyYAML==6.0
render==1.0.0
requests==2.22.0
responses==0.10.16
s3transfer==0.3.7
six==1.12.0
sqlparse==0.2.4
stevedore==3.5.0
tablib==3.1.0
tzdata==2021.5
urllib3==1.25.10
virtualenv==16.7.9
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
